Split GeneratorPluginLoader and GeneratorPluginLocator
