+++ TechTalk.SpecFlow.Generator/Plugins/IGeneratorPluginLocator.cs
@@ -1,0 +1,9 @@
+﻿using TechTalk.SpecFlow.Plugins;
+
+namespace TechTalk.SpecFlow.Generator.Plugins
+{
+    public interface IGeneratorPluginLocator
+    {
+        string LocatePluginAssembly(PluginDescriptor pluginDescriptor);
+    }
+}
