+++ TechTalk.SpecFlow.Generator/TechTalk.SpecFlow.Generator.csproj
@@ -67,8 +67,10 @@
     <Compile Include="Plugins\GeneratorPluginEventArgs.cs" />
     <Compile Include="Plugins\GeneratorPluginEvents.cs" />
     <Compile Include="Plugins\GeneratorPluginLoader.cs" />
+    <Compile Include="Plugins\GeneratorPluginLocator.cs" />
     <Compile Include="Plugins\GeneratorPluginParameters.cs" />
     <Compile Include="Plugins\IGeneratorPlugin.cs" />
+    <Compile Include="Plugins\IGeneratorPluginLocator.cs" />
     <Compile Include="Plugins\IGeneratorPluginLoader.cs" />
     <Compile Include="Project\ISpecFlowProjectReader.cs" />
     <Compile Include="Project\MsBuildProjectReader.cs" />
