+++ Tests/TechTalk.SpecFlow.GeneratorTests/GeneratorTests.csproj
@@ -82,7 +82,7 @@
     <Content Include="Data\CSProj_New\Class1.cs">
       <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
     </Content>
-    <Compile Include="GeneratorPluginLoaderTests.cs" />
+    <Compile Include="GeneratorPluginLocatorTests.cs" />
     <Compile Include="MSBuildRelativePathParserTests.cs" />
     <Compile Include="ScenarioOutlineParserTests.cs" />
     <Compile Include="TestGeneratorTest.cs" />
