+++ changelog.txt
@@ -1,3 +1,16 @@
+2.3.2
+Fixes:
++ Fix issue when using environment variables in a generator plugin's path
+
+API Changes:
++ Generator probing functionality was moved from GeneratorPluginLoader to GeneratorPluginLocator
+
+2.3.1 - 2018-03-01
+Fixes:
++ Fixes stepdefinitionreport to include bindings that use [StepDefinition] https://github.com/techtalk/SpecFlow/pull/1024
++ Fix code-behind generation for SpecFlow+Excel Excel files https://github.com/techtalk/SpecFlow/pull/1048
++ Fix runtime support for MSTest V2 https://github.com/techtalk/SpecFlow/pull/1053
+
 2.3 - 2018-0?-??
 New Features:
 + Expose the current status (result) of the scenario execution in ScenarioContext via ScenarioExecutionStatus property https://github.com/techtalk/SpecFlow/pull/963
